Cleaned up the use of the statType to hash objects around the place.
