+++ src/app/loadout-builder/types.ts
@@ -124,3 +124,15 @@ export const LockableBuckets = {
   classitem: 1585787867,
   ghost: 4023194814,
 };
+
+export const statHashes: { [type in StatTypes]: number } = {
+  Mobility: 2996146975,
+  Resilience: 392767087,
+  Recovery: 1943323491,
+  Discipline: 1735777505,
+  Intellect: 144602215,
+  Strength: 4244567218,
+};
+
+export const statValues = Object.values(statHashes);
+export const statKeys = Object.keys(statHashes) as StatTypes[];
